hackathon_ClaimsAdjudication/requirements.txt
boto3==1.40.50
botocore
bedrock-agentcore==0.1.7
bedrock-agentcore-starter-toolkit==0.1.24
strands-agents
strands-agents[a2a]
strands-agents-tools
pyyaml
ddgs
PyYAML
PyPDF2
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-boto
opentelemetry-sdk-extension-aws
pydantic 
httpx
fastapi
uvicorn[standard]

